requirements.txt
networkx>=3.0
matplotlib>=3.7.0
requests>=2.31.0
openai>=1.3.0
python-dotenv>=1.0.0
fastapi>=0.104.0
beautifulsoup4>=4.12.0
uvicorn>=0.24.0
